Registros de 32 bits

+++ KSP/KSP/support.h
@@ -6,7 +6,7 @@ namespace ksp
 {
 	typedef void* ptr_t;
 	typedef char* stack_ptr_t;
-	typedef uint64_t reg_t;
+	typedef uint32_t reg_t;
 	typedef reg_t* reg_ptr_t;
 
 	typedef uint8_t opcode_t;
